US121479 Refactor for testability
